rename base classes appropriately
