+++ tests/worker/test_model_advancer.py
@@ -1,9 +1,9 @@
 from unittest import TestCase
 
 
-class TestModelAdvancer(TestCase):
+class TestModelAdvancerBase(TestCase):
     def test_init(self):
         # ignore this one for now because of the arg parser
-        # ma = ModelAdvancer()
+        # ma = ModelAdvancerBase()
         # self.assertIsNotNone(ma)
         pass
